refactor: update dashboard and session wallet components for improved user experience
- Enhanced the Dashboard component with new action tabs for deposit and withdraw functionalities, improving accessibility and usability.
- Updated the SessionWallet component to better handle loading states and display balances, ensuring users have clear visibility of their tokens.
- Added new frequency option for DCA settings and refined input handling for frequency selection.
- Improved error and success message displays for better user feedback during transactions.

These changes aim to streamline user interactions and provide a more intuitive interface for managing DCA activities.

+++ src/components/dca/CreateDCAModal.tsx
@@ -217,7 +217,7 @@ export function CreateDCAModal() {
           {/* Frequency */}
           <div>
             <label className="text-label block mb-2">Frequency</label>
-            <Select value={frequencyHours.toString()} onValueChange={(v) => setFrequencyHours(parseInt(v))}>
+            <Select value={frequencyHours.toString()} onValueChange={(v) => setFrequencyHours(parseFloat(v))}>
               <SelectTrigger className="bg-muted border-border">
                 <SelectValue />
               </SelectTrigger>
